Completando el cambio a c_scalar en los solvers

+++ src/test/mpiTest.c
@@ -88,7 +88,7 @@ int main( int argc, char **argv ) {
     	if( writeFlag(&info) ) {
 	    
     	    if(pid == 0) {
-    		printf("Time = %.2f\n", info.time.current);
+    		printf("Time = %.2f\n", (double)info.time.current);
     		printf("Elapsed time = %.2f seconds\n\n", elapsed(&info) );
     	    }
